Ensure login was successful after init;
Beatport seems to respond with 302 when the login was successful and 200
when it was not. The previous implementation expected it to return a 4xx
status code on an unsuccessful login and thus would not return an error
when the login was unsuccessful. It seems that at least at the moment
Beatport returns 403 when making a request to my-beatport without a
valid login session and thus it can be used to see whether the login was
successful or not.